+++ src/net.hpp
@@ -11,20 +11,20 @@ class Net
 {
 private:
     vector<Layer> m_layers; //m_layers[layerNum][neuronNum]
-    float m_error;
-    float m_recentAverageError;
-    static float m_recentAverageSmoothingFactor;
+    double m_error;
+    double m_recentAverageError;
+    static double m_recentAverageSmoothingFactor;
 
 public:
     Net(const vector<unsigned> &topology);
-    void feedForward(const vector<float> &inputVals);
-    void backProp(const vector<float> &targetVals);
+    void feedForward(const vector<double> &inputVals);
+    void backProp(const vector<double> &targetVals);
     void updateWeights();
-    void getResults(vector<float> &resultVals) const;
-    float getRecentAverageError(void) const {return m_recentAverageError;};
-    float getError(void) const {return m_error;};
+    void getResults(vector<double> &resultVals) const;
+    double getRecentAverageError(void) const {return m_recentAverageError;};
+    double getError(void) const {return m_error;};
     void calcAvgGradient(unsigned int batchSize);
     void resetGradientSum();
-    float getLoss(const vector<float> &targetVals);
-    int compare_result(const vector<float> &output, const vector<float> &label);
+    double getLoss(const vector<double> &targetVals);
+    int compare_result(const vector<double> &output, const vector<double> &label);
 };
